ServiceとController部分を実装
