--- src/main/java/raisetech/student/management/system/StudentCourseRepository.java
@@ -1,14 +1,0 @@
-package raisetech.student.management.system;
-
-import java.util.List;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Select;
-
-@Mapper
-public interface StudentCourseRepository {
-
-  @Select("SELECT * FROM students_courses")
-  List<StudentCourse> searchCourse();
-
-
-}
