+++ src/main/java/raisetech/student/management/system/StudentManagementSystemApplication.java
@@ -6,28 +6,18 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import raisetech.student.management.system.data.Student;
+import raisetech.student.management.system.data.StudentCourse;
+import raisetech.student.management.system.repository.StudentRepository;
 
 @SpringBootApplication
 @RestController
 
 public class StudentManagementSystemApplication {
 
-  @Autowired
-  private StudentRepository studentRepository;
-  @Autowired
-  private StudentCourseRepository courseRepository;
 
   public static void main(String[] args) {
     SpringApplication.run(StudentManagementSystemApplication.class, args);
   }
 
-  @GetMapping("/studentList")
-  public List<Student> getStudentList() {
-    return studentRepository.searchStudent();
-  }
-
-  @GetMapping("/courseList")
-  public List<StudentCourse> getCourseList() {
-    return courseRepository.searchCourse();
-  }
 }
