--- src/main/java/raisetech/student/management/system/StudentRepository.java
@@ -1,13 +1,0 @@
-package raisetech.student.management.system;
-
-import java.util.List;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Select;
-
-@Mapper
-public interface StudentRepository {
-
-  @Select("SELECT * FROM students")
-  List<Student> searchStudent();
-
-}
