+++ src/main/java/raisetech/student/management/system/data/Student.java
@@ -1,4 +1,4 @@
-package raisetech.student.management.system;
+package raisetech.student.management.system.data;
 
 import lombok.Getter;
 import lombok.Setter;
